refactor: (gallery/TypographyPage) project structure adjustments
