+++ source/iNKORE.UI.WPF.Modern.Gallery/Controls/UserControls/TypographyControl.xaml
@@ -2,11 +2,11 @@
 <!--  Licensed under the MIT License.  -->
 
 <UserControl
-    x:Class="iNKORE.UI.WPF.Modern.Gallery.Pages.DesignGuidance.TypographyControl"
+    x:Class="iNKORE.UI.WPF.Modern.Gallery.Controls.UserControls.TypographyControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:iNKORE.UI.WPF.Modern.Gallery.Pages.DesignGuidance"
+    xmlns:local="clr-namespace:iNKORE.UI.WPF.Modern.Gallery.Controls.UserControls"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
