+++ source/iNKORE.UI.WPF.Modern.Gallery/DataModel/Data/Controls.Foundation.json
@@ -1,0 +1,71 @@
+{
+    // Everything from Windows UI
+    "$schema": "../ControlInfoData.schema.json",
+
+    "Groups":
+    [
+        {
+            "UniqueId": "NewControls",
+            "Title": "Home",
+            "Subtitle": "",
+            "ImagePath": "",
+            "ImageIconPath": "",
+            "Description": "",
+            "Items":
+            [
+            ]
+        },
+        {
+            "UniqueId": "Design",
+            "Title": "Design",
+            "Subtitle": "",
+            "ImagePath": "",
+            "ImageIconPath": "",
+            "Description": "Design guidance and resources",
+            "Items":
+            [
+                {
+                    "UniqueId": "Typography",
+                    "Title": "Typography",
+                    "Subtitle": "Typography type ramp and system fonts",
+                    "ImagePath": "ms-appx:///Assets/ControlIcons/DefaultIcon.png",
+                    "ImageIconPath": "ms-appx:///Assets/ControlIcons/DefaultIcon.png",
+                    "Description": "",
+                    "Content": "",
+                    "IncludedInBuild": true,
+                    "IsNew": false,
+                    "IsUpdated": false,
+                    "Docs":
+                    [
+                        {
+                            "Title": "Typography in Windows Apps",
+                            "Uri": "https://learn.microsoft.com/windows/apps/design/style/typography"
+                        },
+                        {
+                            "Title": "XAML theme resources",
+                            "Uri": "https://learn.microsoft.com/windows/apps/design/style/xaml-theme-resources#the-xaml-type-ramp"
+                        },
+                        {
+                            "Title": "Typography in Windows 11",
+                            "Uri": "https://learn.microsoft.com/windows/apps/design/signature-experiences/typography"
+                        }
+                    ],
+                    "RelatedControls":
+                    [
+                    ]
+                }
+            ]
+        },
+        {
+            "UniqueId": "AllControls",
+            "Title": "All controls",
+            "Subtitle": "",
+            "ImagePath": "",
+            "ImageIconPath": "",
+            "Description": "",
+            "Items":
+            [
+            ]
+        }
+    ]
+}
